sec update

+++ backend/migrations/20240813211028_parties.js
@@ -4,7 +4,7 @@
  */
 exports.up = function (knex) {
   return knex.schema.createTable("parties", function (table) {
-    table.bigIncrements("party_id"); // BIGINT primary key with auto-increment
+    table.increments("party_id"); // BIGINT primary key with auto-increment
     table.string("party_name", 100); // VARCHAR(100) for the party name
   });
 };
